Lexer bugfixing 2
After further bufixes I approached a problem of newlines and tabs and such special characters. They couldn't be escaped or processed correctly by existing generators so I had to write a special generator just for them. It should work fine now.

Now is a perfect time to write detailed and extensive tests for new modules and the whole grammar package. It's gonna get unmanageable without it.